+++ typescript/src/test/shared/configuration.test.ts
@@ -1,5 +1,5 @@
 import z from 'zod';
-import {isToolAllowed} from '../../shared/configuration';
+import {isToolAllowed} from '@/shared/configuration';
 
 describe('isToolAllowed', () => {
   it('should return true if all permissions are allowed', () => {
