+++ typescript/src/test/shared/customers/functions.test.ts
@@ -1,0 +1,93 @@
+import {createCustomer, listCustomers} from '@/shared/customers/functions';
+
+const Stripe = jest.fn().mockImplementation(() => ({
+  customers: {
+    create: jest.fn(),
+    list: jest.fn(),
+  },
+}));
+
+let stripe: ReturnType<typeof Stripe>;
+
+beforeEach(() => {
+  stripe = new Stripe('fake-api-key');
+});
+
+describe('createCustomer', () => {
+  it('should create a customer and return the id', async () => {
+    const params = {
+      email: 'test@example.com',
+      name: 'Test User',
+    };
+
+    const context = {};
+
+    const mockCustomer = {id: 'cus_123456', email: 'test@example.com'};
+    stripe.customers.create.mockResolvedValue(mockCustomer);
+
+    const result = await createCustomer(stripe, context, params);
+
+    expect(stripe.customers.create).toHaveBeenCalledWith(params, undefined);
+    expect(result).toEqual({id: mockCustomer.id});
+  });
+
+  it('should specify the connected account if included in context', async () => {
+    const params = {
+      email: 'test@example.com',
+      name: 'Test User',
+    };
+
+    const context = {
+      account: 'acct_123456',
+    };
+
+    const mockCustomer = {id: 'cus_123456', email: 'test@example.com'};
+    stripe.customers.create.mockResolvedValue(mockCustomer);
+
+    const result = await createCustomer(stripe, context, params);
+
+    expect(stripe.customers.create).toHaveBeenCalledWith(params, {
+      stripeAccount: context.account,
+    });
+    expect(result).toEqual({id: mockCustomer.id});
+  });
+});
+
+describe('listCustomers', () => {
+  it('should list customers and return their ids', async () => {
+    const mockCustomers = [
+      {id: 'cus_123456', email: 'test1@example.com'},
+      {id: 'cus_789012', email: 'test2@example.com'},
+    ];
+
+    const context = {};
+
+    stripe.customers.list.mockResolvedValue({data: mockCustomers});
+    const result = await listCustomers(stripe, context, {});
+
+    expect(stripe.customers.list).toHaveBeenCalledWith({}, undefined);
+    expect(result).toEqual(mockCustomers.map(({id}) => ({id})));
+  });
+
+  it('should specify the connected account if included in context', async () => {
+    const mockCustomers = [
+      {id: 'cus_123456', email: 'test1@example.com'},
+      {id: 'cus_789012', email: 'test2@example.com'},
+    ];
+
+    const context = {
+      account: 'acct_123456',
+    };
+
+    stripe.customers.list.mockResolvedValue({data: mockCustomers});
+    const result = await listCustomers(stripe, context, {});
+
+    expect(stripe.customers.list).toHaveBeenCalledWith(
+      {},
+      {
+        stripeAccount: context.account,
+      }
+    );
+    expect(result).toEqual(mockCustomers.map(({id}) => ({id})));
+  });
+});
