+++ typescript/src/test/shared/invoices/prompts.test.ts
@@ -2,8 +2,7 @@ import {
   createInvoiceItemPrompt,
   createInvoicePrompt,
   listInvoicesPrompt,
-} from '../../shared/invoices/prompts';
-import {listPaymentIntentsPrompt} from '../../shared/paymentIntents/prompts';
+} from '@/shared/invoices/prompts';
 
 describe('createInvoicePrompt', () => {
   it('should return the correct prompt', () => {
@@ -37,18 +36,6 @@ describe('createInvoiceItemPrompt', () => {
 
   it('should return the correct prompt when a customer is specified', () => {
     const prompt = createInvoiceItemPrompt({customer: 'cus_123'});
-    expect(prompt).not.toContain('customer');
-  });
-});
-
-describe('listPaymentIntentsPrompt', () => {
-  it('should return the correct prompt', () => {
-    const prompt = listPaymentIntentsPrompt();
-    expect(prompt).toContain('customer');
-  });
-
-  it('should return the correct prompt when a customer is specified', () => {
-    const prompt = listPaymentIntentsPrompt({customer: 'cus_123'});
     expect(prompt).not.toContain('customer');
   });
 });
