+++ typescript/src/test/shared/products/functions.test.ts
@@ -1,0 +1,91 @@
+import {createProduct, listProducts} from '@/shared/products/functions';
+
+const Stripe = jest.fn().mockImplementation(() => ({
+  products: {
+    create: jest.fn(),
+    list: jest.fn(),
+  },
+}));
+
+let stripe: ReturnType<typeof Stripe>;
+
+beforeEach(() => {
+  stripe = new Stripe('fake-api-key');
+});
+
+describe('createProduct', () => {
+  it('should create a product and return it', async () => {
+    const params = {
+      name: 'Test Product',
+    };
+
+    const context = {};
+
+    const mockProduct = {id: 'prod_123456', name: 'Test Product'};
+    stripe.products.create.mockResolvedValue(mockProduct);
+
+    const result = await createProduct(stripe, context, params);
+
+    expect(stripe.products.create).toHaveBeenCalledWith(params, undefined);
+    expect(result).toEqual(mockProduct);
+  });
+
+  it('should specify the connected account if included in context', async () => {
+    const params = {
+      name: 'Test Product',
+    };
+
+    const context = {
+      account: 'acct_123456',
+    };
+
+    const mockProduct = {id: 'prod_123456', name: 'Test Product'};
+    stripe.products.create.mockResolvedValue(mockProduct);
+
+    const result = await createProduct(stripe, context, params);
+
+    expect(stripe.products.create).toHaveBeenCalledWith(params, {
+      stripeAccount: context.account,
+    });
+    expect(result).toEqual(mockProduct);
+  });
+});
+
+describe('listProducts', () => {
+  it('should list products and return them', async () => {
+    const mockProducts = [
+      {id: 'prod_123456', name: 'Test Product 1'},
+      {id: 'prod_789012', name: 'Test Product 2'},
+    ];
+
+    const context = {};
+
+    stripe.products.list.mockResolvedValue({data: mockProducts});
+    const result = await listProducts(stripe, context, {});
+
+    expect(stripe.products.list).toHaveBeenCalledWith({}, undefined);
+    expect(result).toEqual(mockProducts);
+  });
+
+  it('should specify the connected account if included in context', async () => {
+    const mockProducts = [
+      {id: 'prod_123456', name: 'Test Product 1'},
+      {id: 'prod_789012', name: 'Test Product 2'},
+    ];
+
+    const context = {
+      account: 'acct_123456',
+    };
+
+    stripe.products.list.mockResolvedValue({data: mockProducts});
+    const result = await listProducts(stripe, context, {});
+
+    expect(stripe.products.list).toHaveBeenCalledWith(
+      {},
+      {
+        stripeAccount: context.account,
+      }
+    );
+    expect(result).toEqual(mockProducts);
+  });
+});
