+++ typescript/src/test/shared/refunds/parameters.test.ts
@@ -1,0 +1,19 @@
+import {createRefundParameters} from '@/shared/refunds/parameters';
+
+describe('createRefundParameters', () => {
+  it('should return the correct parameters if no context', () => {
+    const parameters = createRefundParameters({});
+
+    const fields = Object.keys(parameters.shape);
+    expect(fields).toEqual(['payment_intent', 'amount']);
+    expect(fields.length).toBe(2);
+  });
+
+  it('should return the correct parameters if customer is specified', () => {
+    const parameters = createRefundParameters({customer: 'cus_123'});
+
+    const fields = Object.keys(parameters.shape);
+    expect(fields).toEqual(['payment_intent', 'amount']);
+    expect(fields.length).toBe(2);
+  });
+});
